resolve notes
